Fix entry point in package.json, change directory/file names, and create a script that populates the database
This commit fixes the entry point in package.json to the correct file name and adds a script to populate the database. This commit also changes the name of the config folder to 'db' and 'db.js' to 'pool.js' to better reflect the code written in the file and the overall project structure.