Add authentication using Passport.js
This commit adds authentication to the application by prompting users to log in to their accounts. The POST method for the sign-up form now redirects the user to the login page once they successfully create an account as opposed to the home page. Errors for unsuccessful login attempts are displayed using the connect-flash package, which integrates with Passport's authenticate function.