Display full message details when authenticated user is logged in
This commit updates the index controller to return either partial or full message details depending on whether the user is authenticated or not (req.user) and displays the respective information on the home page. This commit also fixes an error in the getFullMessageDetails query that called the incorrect column name.