Fix internal server error being displayed to the user when signing up
This commit fixes the issue of the internal server error being displayed to the user when submitting a username that already exists. Now, an appropriate error message is displayed on the sign-up page. This commit also prevents the user error message 'The username already exists' from displaying in the console and only displays it to the user on the sign-up page to prevent confusion for developers.