Add a feature that allows the user to create a message
This commit allows an authenticated user (i.e., a 'member' or an 'admin') to create a message and display it on the home page. Users without an account can only see the title and text of the messages, whereas users who are logged in to their accounts can see the full details, including the author and date posted.